checks for null values and if the add or delete was successful

+++ public_html/delCON_route.php
@@ -17,28 +17,36 @@ $result2 = $conn->query($sql);
 $check = $result2->fetch_assoc();
 
 $cid=$data->cid;
+if($cid==''){
+  $stuff[]="Value entered empty. Please re-enter.";
+  echo json_encode($stuff);
 
-
-if($check['TITLE']=='Administrator'||$check['TITLE']=='Mission Leader'){
+}
+else{
+  if($check['TITLE']=='Administrator'||$check['TITLE']=='Mission Leader'){
     $sql= "DELETE FROM CONTRACTOR WHERE CONTRACTOR_ID=$cid;";
 
     $result = $conn->query($sql);
-    if ($result === TRUE) {
-        $stuff[]="Succesfuly deleted!";
-        echo json_encode($stuff);
+    if ($result === TRUE && ($conn->affected_rows > 0)) {
+      $stuff[]="Succesfuly deleted!";
+      echo json_encode($stuff);
     } else {
-        $stuff[]="Error with deletion.";
-        echo json_encode($stuff);
+      $stuff[]="Error with deletion.";
+      echo json_encode($stuff);
     }
 
 
 
-}
-else{
-  $string=["NA"];
+  }
+  else{
+    $string=["NA"];
 
-  echo json_encode($string);
+    echo json_encode($string);
+  }
+
 }
+
+
 $conn->close();
 ?>
