+++ public_html/delPRJ_route.php
@@ -1,29 +1,33 @@
-    <?php
+<?php
 
-    require_once('db_setup.php');
+require_once('db_setup.php');
 
-    $sql = "USE mswanso2_1;";
-    if ($conn->query($sql) === TRUE) {
+$sql = "USE mswanso2_1;";
+if ($conn->query($sql) === TRUE) {
          // echo "using Database mswanso2_1";
-    } else {
+} else {
 	//echo "Error using  database: " . $conn->error;
-    }
+}
 
-    $json=file_get_contents('php://input');
-    $stuff=array();
-    $data=json_decode($json);
-    $sql="SELECT TITLE FROM EMPLOYEE WHERE WORK_ID=".$data->workid;
-    $result2 = $conn->query($sql);
-    $check = $result2->fetch_assoc();
+$json=file_get_contents('php://input');
+$stuff=array();
+$data=json_decode($json);
+$sql="SELECT TITLE FROM EMPLOYEE WHERE WORK_ID=".$data->workid;
+$result2 = $conn->query($sql);
+$check = $result2->fetch_assoc();
 
-    $pid=$data->pid;
+$pid=$data->pid;
+if($pid==''){
+  $stuff[]="Value entered empty. Please re-enter.";
+  echo json_encode($stuff);
 
-
+}
+else{
     if($check['TITLE']=='Administrator'||$check['TITLE']=='Mission Leader'){
         $sql= "DELETE FROM PROJECT WHERE PROJ_ID=$pid;";
 
         $result = $conn->query($sql);
-        if ($result === TRUE) {
+        if ($result === TRUE && ($conn->affected_rows > 0)) {
             $stuff[]="Succesfuly deleted!";
             echo json_encode($stuff);
         } else {
@@ -39,5 +43,10 @@
 
       echo json_encode($string);
   }
-  $conn->close();
-  ?>
+
+}
+
+
+
+$conn->close();
+?>
